Temporarily fix scrolling issue
The last item on the left-side drawer of home page is not visible ostensibly because the header takes up space on the page and causes the main content div to be too large. I am currently unsure of how to properly fix this bug, so for now, I made it easier to overscroll in order to see the last item in the drawer.